scraper/requirements.txt
requests
beautifulsoup4
python-dotenv
cloudscraper

